requirements.txt
python-telegram-bot==21.4
tronpy==0.4.0
python-dotenv==1.0.1
pytest==8.3.2